requirements.txt
ipykernel
python-dotenv
langchain
langchain_groq
langchain-community
transformers
PyPDF
validators==0.28.1
youtube_transcript_api
streamlit
unstructured
pytube
groq